switching to 32-bit-based hash

+++ src/thorin/util/hash.cpp
@@ -3,8 +3,8 @@
 
 namespace thorin {
 
-uint64_t hash(const char* s) {
-    uint64_t seed = thorin::hash_begin();
+uint32_t hash(const char* s) {
+    uint32_t seed = thorin::hash_begin();
     for (const char* p = s; *p != '\0'; ++p)
         seed = thorin::hash_combine(seed, *p);
     return seed;
